Refactor Rejection tables to a new component
